@@ -4,7 +4,7 @@
     "description": "Api backend for OS2display",
     "license": "proprietary",
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "api-platform/core": "^2.6",
@@ -46,12 +46,20 @@
         "ergebnis/composer-normalize": "^2.15",
         "friendsofphp/php-cs-fixer": "^3.0",
         "hautelook/alice-bundle": "^2.9",
+        "justinrainbow/json-schema": "^5.2",
+        "phpunit/phpunit": "^9.5",
+        "psalm/plugin-symfony": "^3.0",
+        "symfony/browser-kit": "5.3.*",
+        "symfony/css-selector": "5.3.*",
         "symfony/debug-bundle": "^5.3",
+        "symfony/http-client": "5.3.*",
         "symfony/maker-bundle": "^1.33",
+        "symfony/phpunit-bridge": "^5.3",
         "symfony/stopwatch": "^5.3",
         "symfony/var-dumper": "^5.3",
         "symfony/web-profiler-bundle": "^5.3",
-        "vimeo/psalm": "^4.8"
+        "vimeo/psalm": "^4.8",
+        "weirdan/doctrine-psalm-plugin": "^1.1"
     },
     "config": {
         "optimize-autoloader": true,
@@ -92,8 +100,17 @@
             "cache:clear": "symfony-cmd",
             "assets:install %PUBLIC_DIR%": "symfony-cmd"
         },
-        "check-coding-standards": [
+        "coding-standards-apply": [
+            "vendor/bin/php-cs-fixer fix"
+        ],
+        "coding-standards-check": [
             "vendor/bin/php-cs-fixer fix --dry-run"
+        ],
+        "tests": [
+            "bin/console --env=test doctrine:database:drop --force --quiet",
+            "bin/console --env=test doctrine:database:create --no-interaction --if-not-exists --quiet",
+            "bin/console --env=test doctrine:migrations:migrate --no-interaction --quiet",
+            "vendor/bin/phpunit --stop-on-failure"
         ]
     }
 }
