@@ -14,23 +14,24 @@ use Gedmo\Timestampable\Traits\TimestampableEntity;
 class Screen
 {
     use EntityIdTrait;
-    use EntityTitleDescTrait;
+    use EntityTitleDescriptionTrait;
+    use EntityModificationTrait;
     use TimestampableEntity;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="integer", options={"default": 0})
      */
-    private $size;
+    private int $size = 0;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="integer", options={"default": 0})
      */
-    private int $resolutionWidth;
+    private int $resolutionWidth = 0;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="integer", options={"default": 0})
      */
-    private int $resolutionHeight;
+    private int $resolutionHeight = 0;
 
     /**
      * @ORM\ManyToOne(targetEntity=ScreenLayout::class, inversedBy="screens")
@@ -39,21 +40,27 @@ class Screen
     private ScreenLayout $screenLayout;
 
     /**
-     * @ORM\Column(type="string", length=255, nullable=true)
+     * @ORM\Column(type="string", length=255, nullable=true, options={"default": ""})
      */
-    private string $location;
+    private string $location = '';
 
     /**
      * @ORM\ManyToMany(targetEntity=Playlist::class, inversedBy="screens")
      */
-    private $playlists;
+    private Collection $playlists;
+
+    /**
+     * @ORM\OneToMany(targetEntity=PlaylistScreenRegion::class, mappedBy="screen", orphanRemoval=true)
+     */
+    private Collection $playlistScreenRegions;
 
     public function __construct()
     {
         $this->playlists = new ArrayCollection();
+        $this->playlistScreenRegions = new ArrayCollection();
     }
 
-    public function getSize(): ?int
+    public function getSize(): int
     {
         return $this->size;
     }
@@ -65,7 +72,7 @@ class Screen
         return $this;
     }
 
-    public function getResolutionWidth(): ?int
+    public function getResolutionWidth(): int
     {
         return $this->resolutionWidth;
     }
@@ -77,7 +84,7 @@ class Screen
         return $this;
     }
 
-    public function getResolutionHeight(): ?int
+    public function getResolutionHeight(): int
     {
         return $this->resolutionHeight;
     }
@@ -89,24 +96,24 @@ class Screen
         return $this;
     }
 
-    public function getScreenLayout(): ?ScreenLayout
+    public function getScreenLayout(): ScreenLayout
     {
         return $this->screenLayout;
     }
 
-    public function setScreenLayout(?ScreenLayout $screenLayout): self
+    public function setScreenLayout(ScreenLayout $screenLayout): self
     {
         $this->screenLayout = $screenLayout;
 
         return $this;
     }
 
-    public function getLocation(): ?string
+    public function getLocation(): string
     {
         return $this->location;
     }
 
-    public function setLocation(?string $location): self
+    public function setLocation(string $location): self
     {
         $this->location = $location;
 
@@ -114,7 +121,7 @@ class Screen
     }
 
     /**
-     * @return Collection|Playlist[]
+     * @return ArrayCollection|Playlist[]
      */
     public function getPlaylists(): Collection
     {
@@ -124,7 +131,7 @@ class Screen
     public function addPlaylist(Playlist $playlist): self
     {
         if (!$this->playlists->contains($playlist)) {
-            $this->playlists[] = $playlist;
+            $this->playlists->add($playlist);
         }
 
         return $this;
@@ -133,6 +140,36 @@ class Screen
     public function removePlaylist(Playlist $playlist): self
     {
         $this->playlists->removeElement($playlist);
+
+        return $this;
+    }
+
+    /**
+     * @return ArrayCollection|PlaylistScreenRegion[]
+     */
+    public function getPlaylistScreenRegions(): Collection
+    {
+        return $this->playlistScreenRegions;
+    }
+
+    public function addPlaylistScreenRegion(PlaylistScreenRegion $playlistScreenRegion): self
+    {
+        if (!$this->playlistScreenRegions->contains($playlistScreenRegion)) {
+            $this->playlistScreenRegions->add($playlistScreenRegion);
+            $playlistScreenRegion->setScreen($this);
+        }
+
+        return $this;
+    }
+
+    public function removePlaylistScreenRegion(PlaylistScreenRegion $playlistScreenRegion): self
+    {
+        if ($this->playlistScreenRegions->removeElement($playlistScreenRegion)) {
+            // set the owning side to null (unless already changed)
+            if ($playlistScreenRegion->getScreen() === $this) {
+                $playlistScreenRegion->removeScreen();
+            }
+        }
 
         return $this;
     }
