@@ -2,48 +2,39 @@
 
 namespace App\Entity;
 
-use DateTimeImmutable;
-use DateTimeInterface;
+use Doctrine\ORM\Mapping as ORM;
 
 trait EntityPublishedTrait
 {
     /**
      * @ORM\Column(type="datetime", nullable=true)
      */
-    private DateTimeImmutable $publishedFrom;
+    private \DateTime $publishedFrom;
 
     /**
      * @ORM\Column(type="datetime", nullable=true)
      */
-    private DateTimeImmutable $publishedTo;
+    private \DateTime $publishedTo;
 
-    public function getPublishedFrom(): DateTimeImmutable
+    public function getPublishedFrom(): \DateTime
     {
         return $this->publishedFrom;
     }
 
-    public function setPublishedFrom(DateTimeInterface $publishedFrom): self
+    public function setPublishedFrom(\DateTime $publishedFrom): self
     {
-        if (!$publishedFrom instanceof DateTimeImmutable) {
-            $publishedFrom = DateTimeImmutable::createFromMutable($publishedFrom);
-        }
-
         $this->publishedFrom = $publishedFrom;
 
         return $this;
     }
 
-    public function getPublishedTo(): DateTimeImmutable
+    public function getPublishedTo(): \DateTime
     {
         return $this->publishedTo;
     }
 
-    public function setPublishedTo(DateTimeInterface $publishedTo): self
+    public function setPublishedTo(\DateTime $publishedTo): self
     {
-        if (!$publishedTo instanceof DateTimeImmutable) {
-            $publishedTo = DateTimeImmutable::createFromMutable($publishedTo);
-        }
-
         $this->publishedTo = $publishedTo;
 
         return $this;
