@@ -2,22 +2,38 @@
 
 namespace App\Entity;
 
-use ApiPlatform\Core\Annotation\ApiResource;
 use App\Repository\MediaRepository;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Timestampable\Traits\TimestampableEntity;
 
 /**
- * @ApiResource()
  * @ORM\Entity(repositoryClass=MediaRepository::class)
  */
 class Media
 {
     use EntityIdTrait;
-    use EntityTitleDescTrait;
     use TimestampableEntity;
+    use EntityTitleDescriptionTrait;
+    use EntityModificationTrait;
 
-    /* TODO Blameable when we have a User entity */
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true, options={"default": ""})
+     */
+    private string $license = '';
 
-    /* TODO Image file handling and upload */
+    // @TODO: Blameable when we have a User entity
+    // @TODO: Image file handling and upload
+    // @TODO: Missing assets (type, uri, dimensions, sha, size)
+
+    public function getLicense(): string
+    {
+        return $this->license;
+    }
+
+    public function setLicense(string $license): self
+    {
+        $this->license = $license;
+
+        return $this;
+    }
 }
