@@ -2,9 +2,15 @@
 
 namespace App\Repository;
 
+use ApiPlatform\Core\Bridge\Doctrine\Orm\Paginator;
+use ApiPlatform\Core\Exception\InvalidArgumentException;
 use App\Entity\Playlist;
+use App\Entity\Slide;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\ORM\Query\Expr\Join;
+use Doctrine\ORM\Tools\Pagination\Paginator as DoctrinePaginator;
 use Doctrine\Persistence\ManagerRegistry;
+use Symfony\Component\Uid\Ulid;
 
 /**
  * @method Playlist|null find($id, $lockMode = null, $lockVersion = null)
@@ -14,37 +20,70 @@ use Doctrine\Persistence\ManagerRegistry;
  */
 class PlaylistRepository extends ServiceEntityRepository
 {
+    public const LINK = 'link';
+    public const UNLINK = 'unlink';
+
     public function __construct(ManagerRegistry $registry)
     {
         parent::__construct($registry, Playlist::class);
     }
 
-    // /**
-    //  * @return Playlist[] Returns an array of Playlist objects
-    //  */
-    /*
-    public function findByExampleField($value)
+    public function getPaginator(string $entityType, Ulid $playlistUid, int $page = 1, int $itemsPerPage = 10): Paginator
     {
-        return $this->createQueryBuilder('p')
-            ->andWhere('p.exampleField = :val')
-            ->setParameter('val', $value)
-            ->orderBy('p.id', 'ASC')
-            ->setMaxResults(10)
-            ->getQuery()
-            ->getResult()
-        ;
-    }
-    */
+        $firstResult = ($page - 1) * $itemsPerPage;
 
-    /*
-    public function findOneBySomeField($value): ?Playlist
-    {
-        return $this->createQueryBuilder('p')
-            ->andWhere('p.exampleField = :val')
-            ->setParameter('val', $value)
-            ->getQuery()
-            ->getOneOrNullResult()
-        ;
+        $queryBuilder = $this->_em->createQueryBuilder();
+        $queryBuilder->select('s')
+            ->from($entityType, 's')
+            ->innerJoin('s.playlists', 'p', Join::WITH, ' p.id = :playlistId')
+            ->setParameter('playlistId', $playlistUid, 'ulid');
+
+        $query = $queryBuilder->getQuery()
+            ->setFirstResult($firstResult)
+            ->setMaxResults($itemsPerPage);
+
+        $doctrinePaginator = new DoctrinePaginator($query);
+
+        return new Paginator($doctrinePaginator);
     }
-    */
+
+    /**
+     * Slide operations (link/unlink slides on a given playlist).
+     *
+     * @param ulid $ulid
+     *   Playlist Ulid for the playlist to manipulate
+     * @param Ulid $slideUlid
+     *   Screen Ulid to link/unlink to the playlist
+     * @param string $op
+     *   The operation to perform (use the PlaylistRepository:: constants)
+     *
+     * @throws \Doctrine\ORM\ORMException
+     * @throws \Doctrine\ORM\OptimisticLockException
+     */
+    public function slideOperation(Ulid $ulid, Ulid $slideUlid, string $op = PlaylistRepository::LINK): void
+    {
+        $slideRepos = $this->getEntityManager()->getRepository(Slide::class);
+        $slide = $slideRepos->findOneBy(['id' => $slideUlid]);
+        if (is_null($slide)) {
+            throw new InvalidArgumentException('Slide not found');
+        }
+
+        $playlistRepos = $this->getEntityManager()->getRepository(Playlist::class);
+        $playlist = $playlistRepos->findOneBy(['id' => $ulid]);
+        if (is_null($playlist)) {
+            throw new InvalidArgumentException('Playlist not found');
+        }
+
+        switch ($op) {
+            case self::LINK:
+                $playlist->addSlide($slide);
+                break;
+
+            case self::UNLINK:
+                $playlist->removeSlide($slide);
+                break;
+        }
+
+        $this->getEntityManager()->flush();
+    }
 }
