@@ -2,9 +2,13 @@
 
 namespace App\Repository;
 
+use ApiPlatform\Core\Bridge\Doctrine\Orm\Paginator;
 use App\Entity\Slide;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\ORM\Query\Expr\Join;
+use Doctrine\ORM\Tools\Pagination\Paginator as DoctrinePaginator;
 use Doctrine\Persistence\ManagerRegistry;
+use Symfony\Component\Uid\Ulid;
 
 /**
  * @method Slide|null find($id, $lockMode = null, $lockVersion = null)
@@ -19,32 +23,22 @@ class SlideRepository extends ServiceEntityRepository
         parent::__construct($registry, Slide::class);
     }
 
-    // /**
-    //  * @return Slide[] Returns an array of Slide objects
-    //  */
-    /*
-    public function findByExampleField($value)
+    public function getPaginator(string $entityType, Ulid $slideUlid, int $page = 1, int $itemsPerPage = 10): Paginator
     {
-        return $this->createQueryBuilder('s')
-            ->andWhere('s.exampleField = :val')
-            ->setParameter('val', $value)
-            ->orderBy('s.id', 'ASC')
-            ->setMaxResults(10)
-            ->getQuery()
-            ->getResult()
-        ;
-    }
-    */
+        $firstResult = ($page - 1) * $itemsPerPage;
 
-    /*
-    public function findOneBySomeField($value): ?Slide
-    {
-        return $this->createQueryBuilder('s')
-            ->andWhere('s.exampleField = :val')
-            ->setParameter('val', $value)
-            ->getQuery()
-            ->getOneOrNullResult()
-        ;
+        $queryBuilder = $this->_em->createQueryBuilder();
+        $queryBuilder->select('p')
+            ->from($entityType, 'p')
+            ->innerJoin('p.slides', 's', Join::WITH, ' s.id = :slideId')
+            ->setParameter('slideId', $slideUlid, 'ulid');
+
+        $query = $queryBuilder->getQuery()
+            ->setFirstResult($firstResult)
+            ->setMaxResults($itemsPerPage);
+
+        $doctrinePaginator = new DoctrinePaginator($query);
+
+        return new Paginator($doctrinePaginator);
     }
-    */
 }
